+++ src/soloud/src/audiosource/wav/dr_impl.cpp
@@ -1,6 +1,8 @@
-#define DR_MP3_IMPLEMENTATION
-#define DR_MP3_NO_STDIO
-#define DR_MP3_FLOAT_OUTPUT
+/// Removing these defines from here to put them in mp3_stream_decoder.cpp because
+/// static members are not used by SoLoud but I need them in mp3_stream_decoder.cpp
+// #define DR_MP3_IMPLEMENTATION
+// #define DR_MP3_NO_STDIO
+// #define DR_MP3_FLOAT_OUTPUT
 #include "dr_mp3.h"
 
 #define DR_WAV_IMPLEMENTATION